stash websocket progress
