+++ client/src/lib.rs
@@ -1,8 +1,20 @@
+
+/// Re-export builder logic
+mod builder;
+pub use crate::builder::*;
+
+/// Re-export client logic
 mod client;
 pub use crate::client::*;
 
+/// Expose database module
 pub mod database;
+
+/// Expose errors module
 pub mod errors;
+
+/// Expose rpc module
 pub mod rpc;
 
+/// Node module is internal to the client crate
 mod node;
